C: Localize token struct members (#529)
## What it does

This PR removes the need for heap allocating the struct members of a
token (position, range, location)

## How it does it

- Change the `token_T` members `range` and `location` to be structs,
instead of pointers to structs
- Change `location_T` members `start` and `end` to be structs, instead
of pointers to structs
- Removes functions only used to access struct members, as they were not
consistently used anyway
- Removes init functions that do not add anything beyond providing a 1-1
mapping of argument to struct members
- Removes copy methods as we are passing the structs by value
- Use 32bit unsigned integers for range/position/location struct
members, effectively limiting the parseable filesize to 2^32-1 bytes
(4gb) which for all intents and purposes (templates after all) should
more than suffice. Saves a bit of memory without any real world
drawbacks.

+++ src/include/lexer_peek_helpers.h
@@ -5,16 +5,17 @@
 #include "token_struct.h"
 
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
 
 typedef struct {
-  size_t position;
-  size_t line;
-  size_t column;
-  size_t previous_position;
-  size_t previous_line;
-  size_t previous_column;
+  uint32_t position;
+  uint32_t line;
+  uint32_t column;
+  uint32_t previous_position;
+  uint32_t previous_line;
+  uint32_t previous_column;
   char current_character;
   lexer_state_T state;
 } lexer_state_snapshot_T;
